Refactor Kubernetes tools to Cgroups tools
