+++ openshift/system/config/unicorn.rb
@@ -6,7 +6,7 @@ require 'etc'
 
 app_path = Pathname.pwd
 require app_path.join('lib/prometheus_exporter_port').to_s
-require app_path.join('lib/kubernetes/tools').to_s
+require app_path.join('lib/cgroups/tools').to_s
 
 Unicorn::HttpServer::START_CTX[0] = '/usr/local/bin/unicorn'
 
@@ -17,13 +17,14 @@ detect_unicorn_workers = -> do
   return Integer(workers) if workers.to_i.positive?
 
   worker_multiplier = Integer(ENV['UNICORN_WORKER_MULTIPLIER'] || 2)
-  cpus = Kubernetes::Tools.kubernetes_cpu_request || Etc.nprocessors
+  cpus = Cgroups::Tools.available_cpus || Etc.nprocessors
 
   return cpus * worker_multiplier
 end
 
-warn "Starting #{detect_unicorn_workers.call} unicorn workers"
-worker_processes detect_unicorn_workers.call
+unicorn_workers_num = detect_unicorn_workers.call
+warn "Starting #{unicorn_workers_num} unicorn workers"
+worker_processes unicorn_workers_num
 
 # listen to the default port
 listen Integer(ENV['PORT'] || 3000)
